+++ src/Polly.Net40/Polly.Net40.csproj
@@ -72,6 +72,9 @@
     <Compile Include="..\Polly.Net35\ExceptionPredicate.cs">
       <Link>ExceptionPredicate.cs</Link>
     </Compile>
+    <Compile Include="..\Polly.Net35\HandledPolicy.cs">
+      <Link>HandledPolicy.cs</Link>
+    </Compile>
     <Compile Include="..\Polly.Net35\HandledPolicySyntax.cs">
       <Link>HandledPolicySyntax.cs</Link>
     </Compile>
