+++ src/Polly.Net45/Polly.Net45.csproj
@@ -71,6 +71,9 @@
     <Compile Include="..\Polly.Net35\ExceptionPredicate.cs">
       <Link>ExceptionPredicate.cs</Link>
     </Compile>
+    <Compile Include="..\Polly.Net35\HandledPolicy.cs">
+      <Link>HandledPolicy.cs</Link>
+    </Compile>
     <Compile Include="..\Polly.Net35\HandledPolicySyntax.cs">
       <Link>HandledPolicySyntax.cs</Link>
     </Compile>
